+++ src/domain/interfaces/IFileImageRepository.ts
@@ -1,9 +1,8 @@
 import {ICrudRepository} from '@steroidsjs/nest/usecases/interfaces/ICrudRepository';
 import {FileImageModel} from '../models/FileImageModel';
-import {FileStorageNameType} from '../types/FileStorageNameType';
 
 export const IFileImageRepository = 'IFileImageRepository';
 
 export interface IFileImageRepository extends ICrudRepository<FileImageModel> {
-    getFilesPathsByStorageName: (storageName: FileStorageNameType) => Promise<string[] | null>,
+    getFilesPathsByStorageName: (storageName: string) => Promise<string[] | null>,
 }