+++ src/domain/interfaces/IFileRepository.ts
@@ -1,12 +1,11 @@
 import {ICrudRepository} from '@steroidsjs/nest/usecases/interfaces/ICrudRepository';
 import {FileModel} from '../models/FileModel';
-import {FileStorageNameType} from '../types/FileStorageNameType';
 
 export const IFileRepository = 'IFileRepository';
 
 export interface IFileRepository extends ICrudRepository<FileModel> {
     getFileWithDocument: (fileName: string) => Promise<FileModel>,
-    getFilesPathsByStorageName: (storageName: FileStorageNameType) => Promise<string[] | null>,
+    getFilesPathsByStorageName: (storageName: string) => Promise<string[] | null>,
     getUnusedFilesIds: (config: {
         fileNameLike: string,
         ignoredTables: string[],