+++ src/domain/interfaces/IFileStorageFactory.ts
@@ -1,8 +1,7 @@
-import {FileStorageNameType} from '../types/FileStorageNameType';
 import {IFileStorage} from './IFileStorage';
 
 export const IFileStorageFactory = 'IFileStorageFactory';
 
 export interface IFileStorageFactory {
-    get: (name: FileStorageNameType) => IFileStorage,
+    get: (name: string) => IFileStorage,
 }