+++ src/domain/types/FileStorageNameType.ts
@@ -1,3 +1,0 @@
-import FileStorageEnum from '../enums/FileStorageEnum';
-
-export type FileStorageNameType = FileStorageEnum | string;