+++ src/usecases/getFilePathModels/interfaces/IGetFileModelsPathUsecase.ts
@@ -1,7 +1,5 @@
-import {FileStorageNameType} from '../../../domain/types/FileStorageNameType';
-
 export const GetFileModelsPathUsecaseToken = 'GetFileModelsPathUsecaseToken';
 
 export interface IGetFileModelsPathUsecase {
-    handle: (storageName: FileStorageNameType) => Promise<string[]>,
+    handle: (storageName: string) => Promise<string[]>,
 }