+++ CHANGELOG.md
@@ -1,14 +1,16 @@
 # Steroids Nest File
 
+## Unreleased
+
 ### Fixes
 
-[Migration guide]()
+[Migration guide](docs/MigrationGuide.md#провайдинг-хранилищ-после-обновления)
 
 - `IFileStorage.write` больше не принимает в качестве аргумента `fileStorageParams`, вместо этого хранилища сами берут эти настройки
 из `IGetFileStorageParamsUseCase`
 - Добавлен токен `FILE_STORAGES_TOKEN`, по которому провайдится объект с хранилищами и их названиями
 - Хранилища теперь имеют injection scope `TRANSIENT`, что позволяет создать несколько экземпляров одного хранилища с разным конфигом
-- Использование `FileStorageEnum` в качестве типа обновлено на `FileStorageNameType`
+- Использование `FileStorageEnum` в качестве типа обновлено на `string`
 
 ## [0.4.1](https://github.com/steroids/nest-file/compare/0.4.0...0.4.1) (2025-12-18)